migration: break down pkg across a few files
We can separate out the `Helper`, `Migration`, and various utilities
into their own files. We'll add tests for individual components in
future commits; the physical separation here sets the foundation for
doing so (prototyped in #57445). This commit is purely code movement.

Release note: None
