+++ pkg/migration/BUILD.bazel
@@ -3,6 +3,7 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library")
 go_library(
     name = "migration",
     srcs = [
+        "helper.go",
         "manager.go",
         "migrations.go",
         "util.go",
